fix: 短箭头 clamp panic、中文 OCR 空格、撤销栈刷屏等 P1/P2 修复 + 规划 M6-M8
P1:
- 箭头 len<20 时 clamp(10, len*0.5) min>max panic（render.rs/canvas.rs），改 max/min
- OCR TSV 中文按词 join(" ") 产生"中 文 识别"，新增 join_words 按 CJK 决定空格
- 粗细滑杆拖动每帧压撤销快照挤爆 100 条上限，只在 drag_started 压一次
- setup_fonts 把未校验字体字节交给 epaint（解析失败 panic!），先 ab_glyph 试解析
- 马赛克缓存只比点数，移动后预览不跟随导出，缓存键改几何指纹哈希
- 改样式不重测 Text 包围盒，命中框与渲染错位

P2:
- 选区拖到屏幕边缘被压扁，改位移预夹取
- 空选区导出静默变整屏（泄漏选区外内容），crop_rgba 改 Option 显式报错
- --duration inf/NaN 触发 Duration::from_secs_f32 panic
- parse_region 放过 NaN；capture_region x+w 溢出改 saturating_add
- w*h*4 在 u32 域回绕，改 usize 域计算；average_color 下标改 get()

docs: PLAN.md 定位从"用完即走"改为"小而美常驻"（托盘空闲<30MB 可度量目标），
新增 M6 图标化工具栏 / M7 贴图 / M8 托盘+配置面板规划；README 同步

+++ crates/core/src/qr.rs
@@ -6,7 +6,7 @@ pub struct QrResult {
 
 /// 在 RGBA 图像中检测并解码所有二维码。
 pub fn detect(rgba: &[u8], w: u32, h: u32) -> Vec<QrResult> {
-    if rgba.len() < (w * h * 4) as usize || w == 0 || h == 0 {
+    if rgba.len() < (w as usize) * (h as usize) * 4 || w == 0 || h == 0 {
         return Vec::new();
     }
     let (w, h) = (w as usize, h as usize);
